chore: refactor archive.rs to eliminate duplication and add schema dump
This commit refactors the archive command to use a trait-based approach
that eliminates code duplication across the four table types (flights,
fixes, receiver_statuses, aprs_messages). It also adds automatic schema
dumping to schema.sql after successful migrations in non-production
environments.

## Archive Refactoring

- Create src/commands/archive/ directory structure
- Extract generic Archivable trait in archiver.rs
- Implement trait for each table type in separate files
- Reduce duplication from ~1,200 lines to ~700 lines total
- Maintain identical CLI interface and functionality

## Schema Dump Feature

- Add dump_schema_if_non_production() function in main.rs
- Automatically runs after successful database migrations
- Only runs when SOAR_ENV != "production"
- Uses pg_dump with deterministic flags (--no-owner, --no-privileges,
  --no-tablespaces, --no-comments) to ensure reproducible output
- Outputs to schema.sql in repository root

## Other Changes

- Add Insertable derive to AprsMessage for archive resurrection support
- Add Context import to main.rs for error handling
